- adjusted the Guestbook, News and Newsletter modules
