+++ ACP3/Modules/ACP3/News/Resources/config/services.yml
@@ -7,7 +7,8 @@ parameters:
   news.controller.sidebar.index.fqdn: 'ACP3\Modules\ACP3\News\Controller\Sidebar\Index'
   news.cache.fqdn: 'ACP3\Modules\ACP3\News\Cache'
   news.model.newsrepository.fqdn: 'ACP3\Modules\ACP3\News\Model\NewsRepository'
-  news.validator.fqdn: 'ACP3\Modules\ACP3\News\Validation\FormValidation'
+  news.validator.fqdn: 'ACP3\Modules\ACP3\News\Validation\AdminFormValidation'
+  news.validation.admin_settings_form_validation.fqdn: 'ACP3\Modules\ACP3\News\Validation\AdminSettingsFormValidation'
   news.validation.validation_rules.external_link_validation_rule.fqdn: 'ACP3\Modules\ACP3\News\Validation\ValidationRules\ExternalLinkValidationRule'
   news.installer.schema.fqdn: 'ACP3\Modules\ACP3\News\Installer\Schema'
   news.installer.migration.fqdn: 'ACP3\Modules\ACP3\News\Installer\Migration'
@@ -16,13 +17,27 @@ parameters:
 services:
   news.controller.admin.index:
     class: %news.controller.admin.index.fqdn%
-    arguments: ['@core.context.admin', '@core.date', '@core.helpers.formToken', '@news.model.newsrepository', '@news.cache', '@news.validator', '@categories.helpers']
+    arguments:
+      - @core.context.admin
+      - @core.date
+      - @core.helpers.formToken
+      - @news.model.newsrepository
+      - @news.cache
+      - @news.validator
+      - @categories.helpers
     calls:
       - ['setCommentsHelpers', ['@?comments.helpers']]
 
   news.controller.frontend.index:
     class: %news.controller.frontend.index.fqdn%
-    arguments: ['@core.context.frontend', '@core.date', '@core.pagination', '@news.model.newsrepository', '@news.cache', '@categories.helpers', '@categories.model']
+    arguments:
+      - @core.context.frontend
+      - @core.date
+      - @core.pagination
+      - @news.model.newsrepository
+      - @news.cache
+      - @categories.helpers
+      - @categories.model
 
   news.controller.sidebar.index:
     class: %news.controller.sidebar.index.fqdn%
@@ -42,7 +57,14 @@ services:
 
   news.validator:
     class: %news.validator.fqdn%
-    arguments: ['@core.lang', @core.validator, '@core.modules']
+    parent: core.validator.abstract
+
+  news.validation.admin_settings_form_validation:
+    class: %news.validation.admin_settings_form_validation.fqdn%
+    arguments:
+      - @core.lang
+      - @core.validator
+      - @core.modules
 
   news.validation.validation_rules.external_link_validation_rule:
     class: %news.validation.validation_rules.external_link_validation_rule.fqdn%
