+++ ACP3/Modules/ACP3/News/Validation/AdminFormValidation.php
@@ -1,0 +1,84 @@
+<?php
+namespace ACP3\Modules\ACP3\News\Validation;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Categories;
+use ACP3\Modules\ACP3\News\Validation\ValidationRules\ExternalLinkValidationRule;
+use ACP3\Modules\ACP3\Seo\Validation\ValidationRules\UriAliasValidationRule;
+
+/**
+ * Class AdminFormValidation
+ * @package ACP3\Modules\ACP3\News\Validation
+ */
+class AdminFormValidation extends Core\Validation\AbstractFormValidation
+{
+    protected $uriAlias = '';
+
+    /**
+     * @param string $uriAlias
+     *
+     * @return $this
+     */
+    public function setUriAlias($uriAlias)
+    {
+        $this->uriAlias = $uriAlias;
+
+        return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function validate(array $formData)
+    {
+        $this->validator
+            ->addConstraint(Core\Validation\ValidationRules\FormTokenValidationRule::NAME)
+            ->addConstraint(
+                Core\Validation\ValidationRules\DateValidationRule::NAME,
+                [
+                    'data' => $formData,
+                    'field' => ['start', 'end'],
+                    'message' => $this->translator->t('system', 'select_date')
+                ])
+            ->addConstraint(
+                Core\Validation\ValidationRules\NotEmptyValidationRule::NAME,
+                [
+                    'data' => $formData,
+                    'field' => 'title',
+                    'message' => $this->translator->t('news', 'title_to_short')
+                ])
+            ->addConstraint(
+                Core\Validation\ValidationRules\NotEmptyValidationRule::NAME,
+                [
+                    'data' => $formData,
+                    'field' => 'text',
+                    'message' => $this->translator->t('news', 'text_to_short')
+                ])
+            ->addConstraint(
+                Categories\Validation\ValidationRules\CategoryExistsValidationRule::NAME,
+                [
+                    'data' => $formData,
+                    'field' => ['cat', 'cat_create'],
+                    'message' => $this->translator->t('news', 'select_category')
+                ])
+            ->addConstraint(
+                ExternalLinkValidationRule::NAME,
+                [
+                    'data' => $formData,
+                    'field' => ['link_title', 'uri', 'target'],
+                    'message' => $this->translator->t('news', 'complete_hyperlink_statements')
+                ])
+            ->addConstraint(
+                UriAliasValidationRule::NAME,
+                [
+                    'data' => $formData,
+                    'field' => 'alias',
+                    'message' => $this->translator->t('seo', 'alias_unallowed_characters_or_exists'),
+                    'extra' => [
+                        'path' => $this->uriAlias
+                    ]
+                ]);
+
+        $this->validator->validate();
+    }
+}
