+++ ACP3/Modules/ACP3/News/Validation/AdminSettingsFormValidation.php
@@ -2,15 +2,13 @@
 namespace ACP3\Modules\ACP3\News\Validation;
 
 use ACP3\Core;
-use ACP3\Modules\ACP3\Categories;
-use ACP3\Modules\ACP3\News\Validation\ValidationRules\ExternalLinkValidationRule;
-use ACP3\Modules\ACP3\Seo\Validation\ValidationRules\UriAliasValidationRule;
+use ACP3\Core\Validation\AbstractFormValidation;
 
 /**
- * Class Validator
- * @package ACP3\Modules\ACP3\News
+ * Class AdminSettingsFormValidation
+ * @package ACP3\Modules\ACP3\News\Validation
  */
-class FormValidation extends Core\Validation\AbstractFormValidation
+class AdminSettingsFormValidation extends AbstractFormValidation
 {
     /**
      * @var Core\Modules
@@ -28,80 +26,16 @@ class FormValidation extends Core\Validation\AbstractFormValidation
         Core\I18n\Translator $translator,
         Core\Validation\Validator $validator,
         Core\Modules $modules
-    )
-    {
+    ) {
         parent::__construct($translator, $validator);
 
         $this->modules = $modules;
     }
 
     /**
-     * @param array $formData
-     *
-     * @throws \ACP3\Core\Exceptions\ValidationFailed
-     * @internal param string $uriAlias
-     *
+     * @inheritdoc
      */
     public function validate(array $formData)
-    {
-        $this->validator
-            ->addConstraint(Core\Validation\ValidationRules\FormTokenValidationRule::NAME)
-            ->addConstraint(
-                Core\Validation\ValidationRules\DateValidationRule::NAME,
-                [
-                    'data' => $formData,
-                    'field' => ['start', 'end'],
-                    'message' => $this->translator->t('system', 'select_date')
-                ])
-            ->addConstraint(
-                Core\Validation\ValidationRules\NotEmptyValidationRule::NAME,
-                [
-                    'data' => $formData,
-                    'field' => 'title',
-                    'message' => $this->translator->t('news', 'title_to_short')
-                ])
-            ->addConstraint(
-                Core\Validation\ValidationRules\NotEmptyValidationRule::NAME,
-                [
-                    'data' => $formData,
-                    'field' => 'text',
-                    'message' => $this->translator->t('news', 'text_to_short')
-                ])
-            ->addConstraint(
-                Categories\Validation\ValidationRules\CategoryExistsValidationRule::NAME,
-                [
-                    'data' => $formData,
-                    'field' => ['cat', 'cat_create'],
-                    'message' => $this->translator->t('news', 'select_category')
-                ])
-            ->addConstraint(
-                ExternalLinkValidationRule::NAME,
-                [
-                    'data' => $formData,
-                    'field' => ['link_title', 'uri', 'target'],
-                    'message' => $this->translator->t('news', 'complete_hyperlink_statements')
-                ])
-            ->addConstraint(
-                UriAliasValidationRule::NAME,
-                [
-                    'data' => $formData,
-                    'field' => 'alias',
-                    'message' => $this->translator->t('seo', 'alias_unallowed_characters_or_exists'),
-                    'extra' => [
-                        'path' => $uriAlias
-                    ]
-                ]);
-
-        $this->validator->validate();
-    }
-
-    /**
-     * @param array $formData
-     *
-     * @throws Core\Exceptions\InvalidFormToken
-     * @throws Core\Exceptions\ValidationFailed
-     */
-    public function validateSettings(array $formData)
     {
         $this->validator
             ->addConstraint(Core\Validation\ValidationRules\FormTokenValidationRule::NAME)
