+++ ACP3/Modules/ACP3/Newsletter/Validation/AdminSettingsFormValidation.php
@@ -1,0 +1,32 @@
+<?php
+namespace ACP3\Modules\ACP3\Newsletter\Validation;
+
+
+use ACP3\Core;
+use ACP3\Core\Validation\AbstractFormValidation;
+
+/**
+ * Class AdminSettingsFormValidation
+ * @package ACP3\Modules\ACP3\Newsletter\Validation
+ */
+class AdminSettingsFormValidation extends AbstractFormValidation
+{
+
+    /**
+     * @inheritdoc
+     */
+    public function validate(array $formData)
+    {
+        $this->validator
+            ->addConstraint(Core\Validation\ValidationRules\FormTokenValidationRule::NAME)
+            ->addConstraint(
+                Core\Validation\ValidationRules\EmailValidationRule::NAME,
+                [
+                    'data' => $formData,
+                    'field' => 'mail',
+                    'message' => $this->translator->t('system', 'wrong_email_format')
+                ]);
+
+        $this->validator->validate();
+    }
+}
